+++ lib/xlsx/xform/drawing/one-cell-anchor-xform.js
@@ -3,8 +3,8 @@ const StaticXform = require('../static-xform');
 
 const CellPositionXform = require('./cell-position-xform');
 const ExtXform = require('./ext-xform');
-const PicXform = require('./pic-xform');
-const SpXform = require('./sp-xform');
+const PicXform = require('./picture/pic-xform');
+const SpXform = require('./shape/sp-xform');
 
 class OneCellAnchorXform extends BaseCellAnchorXform {
   constructor() {
