+++ lib/xlsx/xform/drawing/picture/c-nv-pic-pr-xform.js
@@ -1,4 +1,4 @@
-const BaseXform = require('../base-xform');
+const BaseXform = require('../../base-xform');
 
 class CNvPicPrXform extends BaseXform {
   get tag() {
