+++ lib/xlsx/xform/drawing/picture/nv-pic-pr-xform.js
@@ -1,26 +1,25 @@
-const BaseXform = require('../base-xform');
-const XfrmXform = require('./xfrm-xform');
-const PrstGeomXform = require('./prst-geom-xform');
+const BaseXform = require('../../base-xform');
+const CNvPrXform = require('../c-nv-pr-xform');
+const CNvPicPrXform = require('./c-nv-pic-pr-xform');
 
-// DocumentFormat.OpenXml.Drawing.Spreadsheet.ShapeProperties
-class SpPrXform extends BaseXform {
+class NvPicPrXform extends BaseXform {
   constructor() {
     super();
 
     this.map = {
-      'a:xfrm': new XfrmXform(),
-      'a:prstGeom': new PrstGeomXform(),
+      'xdr:cNvPr': new CNvPrXform(true),
+      'xdr:cNvPicPr': new CNvPicPrXform(),
     };
   }
 
   get tag() {
-    return 'xdr:spPr';
+    return 'xdr:nvPicPr';
   }
 
-  render(xmlStream, shape) {
+  render(xmlStream, model) {
     xmlStream.openNode(this.tag);
-    this.map['a:xfrm'].render(xmlStream, shape);
-    this.map['a:prstGeom'].render(xmlStream, shape);
+    this.map['xdr:cNvPr'].render(xmlStream, model);
+    this.map['xdr:cNvPicPr'].render(xmlStream, model);
     xmlStream.closeNode();
   }
 
@@ -32,7 +31,7 @@ class SpPrXform extends BaseXform {
 
     switch (node.name) {
       case this.tag:
-        this.model = {};
+        this.reset();
         break;
       default:
         this.parser = this.map[node.name];
@@ -55,10 +54,7 @@ class SpPrXform extends BaseXform {
     }
     switch (name) {
       case this.tag:
-        this.mergeModel({
-          ...this.map['a:xfrm'].model,
-          ...this.map['a:prstGeom'].model,
-        });
+        this.model = this.map['xdr:cNvPr'].model;
         return false;
       default:
         return true;
@@ -66,4 +62,4 @@ class SpPrXform extends BaseXform {
   }
 }
 
-module.exports = SpPrXform;
+module.exports = NvPicPrXform;
