+++ lib/xlsx/xform/drawing/picture/pic-xform.js
@@ -1,10 +1,10 @@
-const BaseXform = require('../base-xform');
-const StaticXform = require('../static-xform');
+const BaseXform = require('../../base-xform');
+const StaticXform = require('../../static-xform');
 
-const BlipFillXform = require('./blip-fill-xform');
+const BlipFillXform = require('../blip-fill-xform');
 const NvPicPrXform = require('./nv-pic-pr-xform');
 
-const spPrJSON = require('./sp-pr');
+const spPrJSON = require('../sp-pr');
 
 class PicXform extends BaseXform {
   constructor() {
