+++ lib/xlsx/xform/drawing/shape/nv-sp-pr-xform.js
@@ -1,20 +1,24 @@
-const BaseXform = require('../base-xform');
+const BaseXform = require('../../base-xform');
+const CNvPrXform = require('../c-nv-pr-xform');
 
-// DocumentFormat.OpenXml.Drawing.PresetGeometry
-class PrstGeomXform extends BaseXform {
+// DocumentFormat.OpenXml.Drawing.Spreadsheet.NonVisualShapeProperties
+class NvSpPrXform extends BaseXform {
   constructor() {
     super();
 
-    this.map = {};
+    this.map = {
+      'xdr:cNvPr': new CNvPrXform(false),
+    };
   }
 
   get tag() {
-    return 'a:prstGeom';
+    return 'xdr:nvSpPr';
   }
 
-  render(xmlStream, model) {
-    xmlStream.openNode(this.tag, {prst: model.type});
-    xmlStream.leafNode('a:avLst', {});
+  render(xmlStream, shape) {
+    xmlStream.openNode(this.tag);
+    this.map['xdr:cNvPr'].render(xmlStream, shape);
+    xmlStream.leafNode('xdr:cNvSpPr');
     xmlStream.closeNode();
   }
 
@@ -26,7 +30,6 @@ class PrstGeomXform extends BaseXform {
 
     switch (node.name) {
       case this.tag:
-        this.model = {type: node.attributes.prst};
         break;
       default:
         this.parser = this.map[node.name];
@@ -49,7 +52,7 @@ class PrstGeomXform extends BaseXform {
     }
     switch (name) {
       case this.tag:
-        // TOOD avLst
+        this.model = this.map['xdr:cNvPr'].model;
         return false;
       default:
         return true;
@@ -57,4 +60,4 @@ class PrstGeomXform extends BaseXform {
   }
 }
 
-module.exports = PrstGeomXform;
+module.exports = NvSpPrXform;
