+++ lib/xlsx/xform/drawing/shape/prst-geom-xform.js
@@ -1,24 +1,20 @@
-const BaseXform = require('../base-xform');
-const CNvPrXform = require('./c-nv-pr-xform');
+const BaseXform = require('../../base-xform');
 
-// DocumentFormat.OpenXml.Drawing.Spreadsheet.NonVisualShapeProperties
-class NvSpPrXform extends BaseXform {
+// DocumentFormat.OpenXml.Drawing.PresetGeometry
+class PrstGeomXform extends BaseXform {
   constructor() {
     super();
 
-    this.map = {
-      'xdr:cNvPr': new CNvPrXform(false),
-    };
+    this.map = {};
   }
 
   get tag() {
-    return 'xdr:nvSpPr';
+    return 'a:prstGeom';
   }
 
-  render(xmlStream, shape) {
-    xmlStream.openNode(this.tag);
-    this.map['xdr:cNvPr'].render(xmlStream, shape);
-    xmlStream.leafNode('xdr:cNvSpPr');
+  render(xmlStream, model) {
+    xmlStream.openNode(this.tag, {prst: model.type});
+    xmlStream.leafNode('a:avLst', {});
     xmlStream.closeNode();
   }
 
@@ -30,6 +26,7 @@ class NvSpPrXform extends BaseXform {
 
     switch (node.name) {
       case this.tag:
+        this.model = {type: node.attributes.prst};
         break;
       default:
         this.parser = this.map[node.name];
@@ -52,7 +49,7 @@ class NvSpPrXform extends BaseXform {
     }
     switch (name) {
       case this.tag:
-        this.model = this.map['xdr:cNvPr'].model;
+        // TOOD avLst
         return false;
       default:
         return true;
@@ -60,4 +57,4 @@ class NvSpPrXform extends BaseXform {
   }
 }
 
-module.exports = NvSpPrXform;
+module.exports = PrstGeomXform;
