+++ lib/xlsx/xform/drawing/shape/sp-pr-xform.js
@@ -1,25 +1,26 @@
-const BaseXform = require('../base-xform');
-const CNvPrXform = require('./c-nv-pr-xform');
-const CNvPicPrXform = require('./c-nv-pic-pr-xform');
+const BaseXform = require('../../base-xform');
+const XfrmXform = require('../xfrm-xform');
+const PrstGeomXform = require('./prst-geom-xform');
 
-class NvPicPrXform extends BaseXform {
+// DocumentFormat.OpenXml.Drawing.Spreadsheet.ShapeProperties
+class SpPrXform extends BaseXform {
   constructor() {
     super();
 
     this.map = {
-      'xdr:cNvPr': new CNvPrXform(true),
-      'xdr:cNvPicPr': new CNvPicPrXform(),
+      'a:xfrm': new XfrmXform(),
+      'a:prstGeom': new PrstGeomXform(),
     };
   }
 
   get tag() {
-    return 'xdr:nvPicPr';
+    return 'xdr:spPr';
   }
 
-  render(xmlStream, model) {
+  render(xmlStream, shape) {
     xmlStream.openNode(this.tag);
-    this.map['xdr:cNvPr'].render(xmlStream, model);
-    this.map['xdr:cNvPicPr'].render(xmlStream, model);
+    this.map['a:xfrm'].render(xmlStream, shape);
+    this.map['a:prstGeom'].render(xmlStream, shape);
     xmlStream.closeNode();
   }
 
@@ -31,7 +32,7 @@ class NvPicPrXform extends BaseXform {
 
     switch (node.name) {
       case this.tag:
-        this.reset();
+        this.model = {};
         break;
       default:
         this.parser = this.map[node.name];
@@ -54,7 +55,10 @@ class NvPicPrXform extends BaseXform {
     }
     switch (name) {
       case this.tag:
-        this.model = this.map['xdr:cNvPr'].model;
+        this.mergeModel({
+          ...this.map['a:xfrm'].model,
+          ...this.map['a:prstGeom'].model,
+        });
         return false;
       default:
         return true;
@@ -62,4 +66,4 @@ class NvPicPrXform extends BaseXform {
   }
 }
 
-module.exports = NvPicPrXform;
+module.exports = SpPrXform;
