+++ lib/xlsx/xform/drawing/shape/sp-xform.js
@@ -1,5 +1,5 @@
-const BaseXform = require('../base-xform');
-const StaticXform = require('../static-xform');
+const BaseXform = require('../../base-xform');
+const StaticXform = require('../../static-xform');
 
 const NvSpPrXform = require('./nv-sp-pr-xform');
 const SpPrXform = require('./sp-pr-xform');
