+++ lib/xlsx/xform/drawing/two-cell-anchor-xform.js
@@ -2,8 +2,8 @@ const BaseCellAnchorXform = require('./base-cell-anchor-xform');
 const StaticXform = require('../static-xform');
 
 const CellPositionXform = require('./cell-position-xform');
-const PicXform = require('./pic-xform');
-const SpXform = require('./sp-xform');
+const PicXform = require('./picture/pic-xform');
+const SpXform = require('./shape/sp-xform');
 
 class TwoCellAnchorXform extends BaseCellAnchorXform {
   constructor() {
